feat: 기초 셋팅
폴더위치 수정 및 관리자 전용 controller & service생성 작업을 진행하였습니다.
