+++ src/main/java/com/github/backendpart/service/product/AdminProductService.java
@@ -1,71 +1,33 @@
-package com.github.backendpart.service;
+package com.github.backendpart.service.product;
 
 import com.amazonaws.services.kms.model.NotFoundException;
 import com.github.backendpart.repository.CategoryRepository;
 import com.github.backendpart.repository.ProductRepository;
+import com.github.backendpart.service.ImageUploadService;
+import com.github.backendpart.service.OptionService;
 import com.github.backendpart.web.dto.common.CommonResponseDto;
 import com.github.backendpart.web.dto.product.addProduct.AddProductRequestDto;
-import com.github.backendpart.web.dto.product.OptionDto;
-import com.github.backendpart.web.dto.product.ProductDto;
-import com.github.backendpart.web.dto.product.ProductImageDto;
-import com.github.backendpart.web.dto.product.getProduct.GetProductDetailResponseDto;
-import com.github.backendpart.web.dto.product.getProduct.GetProductResponseDto;
 import com.github.backendpart.web.entity.CategoryEntity;
 import com.github.backendpart.web.entity.OptionEntity;
 import com.github.backendpart.web.entity.ProductEntity;
 import com.github.backendpart.web.entity.ProductImageEntity;
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class ProductService {
+public class AdminProductService {
     private final CategoryRepository categoryRepository;
     private final ProductRepository productRepository;
 
     private final ImageUploadService imageUploadService;
     private final OptionService optionService;
-    
-    public GetProductDetailResponseDto findById(Long productId){
-        ProductEntity targetProduct = productRepository.findByProductCid(productId).orElseThrow(() -> new IllegalArgumentException("해당 Product가 없습니다"));
 
-        //판매기간만 조회 가능하도록
-        if(targetProduct.isSellable()) {
-            return GetProductDetailResponseDto.toDto(targetProduct);
-        }
-
-        return null;
-    }
-
-    @Transactional
-    public List<GetProductResponseDto> findByCategory(String categoryName){
-        log.info("[GetProduct] 요청이 들어왔습니다");
-        List<ProductEntity> targetProductsEntity = productRepository.findByCategory_CategoryName(categoryName);
-        List<GetProductResponseDto> targetProductsDto = new ArrayList<>();
-
-        if(targetProductsEntity == null){
-            log.info("[GetProduct] 해당 카테고리에 상품이 없습니다");
-        }
-        else {
-            for (ProductEntity productEntity : targetProductsEntity) {
-                if(productEntity.isSellable()){
-                    targetProductsDto.add(GetProductResponseDto.toDto(productEntity));
-                    log.info("[Product] 카테고리 조회 리스트에 추가된 상품 : " + productEntity.getProductName());
-                }
-            }
-        }
-
-        return targetProductsDto;
-    }
-    
-    
     public CommonResponseDto addProduct(AddProductRequestDto addProductRequestDto, List<MultipartFile> images){
         try {
             log.info("[addProduct] 새로운 상품 추가 요청이 들어왔습니다. addProductRequestDto = " + addProductRequestDto);
@@ -118,7 +80,7 @@ public class ProductService {
 
         try{
             for(Long targetCid: productCidList){
-            productRepository.deleteById(targetCid);
+                productRepository.deleteById(targetCid);
             }
             return CommonResponseDto.builder()
                     .code(200)
@@ -134,4 +96,6 @@ public class ProductService {
                     .build();
         }
     }
+
+
 }
